Filter ineligible sundae from the total, so we compute the percentile correctly
Usually there's less than 0.5% of the Delegated SUNDAE undelegated/disqualified, and so this hasn't had an effect. But right now, there's 20m SUNDAE delegated to a disqualified pool, and so it caused the percentile computation to be incorrect. This filteres it out of the total in addition to the pool selection, so the percentiles will be correctly calculated